Renaming precondition constants class, fixing preconditions-Book model
