+++ src/main/java/wolox/training/constants/PreconditionsConstants.java
@@ -1,9 +1,10 @@
 package wolox.training.constants;
 
-public class PreconditionsMessages {
+public class PreconditionsConstants {
   public static final String CANNOT_BE_NULL = "Argument %s cannot be null.";
   public static final String CANNOT_BE_EMPTY = "Argument %s cannot be empty.";
   public static final String UNBORN = "User must be born already. Provided date: %s";
   public static final String PAGES_MUST_BE_POSITIVE_NUMBER = "Book pages must be a positive number. Pages number provided: %s";
+  public static final String INVALID_YEAR = "Year must be positive and not greater than actual year.";
 
 }
